refactor: добавлены классы для генерации Id, добавлены константы для валидации

+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -1,13 +1,11 @@
 package ru.yandex.practicum.filmorate.model;
 
 import lombok.Data;
-import lombok.EqualsAndHashCode;
-
 import java.time.LocalDate;
 
 @Data
 public class User {
-    private int id;
+    private long id;
     private String email;
     private String login;
     private String name;
